Show who made each edit on public body admin pages
We already store `PublicBody::Version#last_edit_editor`, so this just
presents it alongside the relevant comment related to that version.

I've handled the case where we can't find a `User` record for an editor
since a `User#url_name` may change if an admin requests anonymisation
and then the link would break. Ideally we'd use an association between
the records rather than having to do a lookup based on a `String`, but
this adds some value now with what we have.

Fixes https://github.com/mysociety/alaveteli/issues/6556